init only runs on initial 'vagrant up'

+++ env/provision.sh
@@ -3,14 +3,28 @@
 set -e
 set -x
 
-sudo add-apt-repository -y ppa:ubuntu-toolchain-r/test && \
-    sudo apt-get update -qq
-sudo apt-get install -qq -y g{cc,++}-4.8 valgrind libboost1.48-all-dev
-sudo pip install --upgrade pip cpp-coveralls
-sudo update-alternatives --install /usr/bin/g++ g++ /usr/bin/g++-4.8 90
-sudo update-alternatives --install /usr/bin/gcov gcov /usr/bin/gcov-4.8 90
+apt='sudo apt-get -qq -y'
+updated_flag=.updated_apt_get
+installed_flag=.installed_dependencies
 
-cat <<EOF >> /home/vagrant/.bashrc
+if ! test -f ${updated_flag}; then
+    touch ${updated_flag}
+    ${apt} update
+fi
+
+if ! test -f ${installed_flag} > /dev/null; then
+    touch ${installed_flag}
+    ${apt} install  curl git python-software-properties
+    ${apt} install  {c,}make build-essential python-{pip,dev}
+    sudo add-apt-repository -y ppa:ubuntu-toolchain-r/test && \
+        ${apt} update
+    ${apt} install  g{cc,++}-4.8 valgrind libboost1.48-all-dev
+    sudo pip install --upgrade pip cpp-coveralls
+    sudo update-alternatives --install /usr/bin/g++ g++ /usr/bin/g++-4.8 90
+    sudo update-alternatives --install /usr/bin/gcov gcov /usr/bin/gcov-4.8 90
+    cat <<EOF >> /home/vagrant/.bashrc
 export CMAKE_ARGS="-DPROFILE=1"
 source /vagrant/configure
 EOF
+fi
+
